builder/hyperv-vmcx fix required properties, add unit tests
